% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini.R
\name{gemini}
\alias{gemini}
\title{Submit a prompt to gemini}
\usage{
gemini(prompt, model = "2.5-flash", image)
}
\arguments{
\item{prompt}{Text prompt}

\item{model}{Optional Gemini model name}

\item{image}{Optional image file}
}
\description{
Submit a prompt to gemini
}
\examples{
\dontrun{
gemini("Qual a capital de Portugal?")
}
}
