% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmstudio.R
\name{lmstudio}
\alias{lmstudio}
\title{Submit a prompt to local LM Studio LLM}
\usage{
lmstudio(prompt)
}
\arguments{
\item{prompt}{Text prompt}
}
\description{
Submit a prompt to local LM Studio LLM
}
\examples{
\dontrun{
lmstudio("Qual a capital de Portugal?")
}
}
