% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmstudio_image.R
\name{lmstudio_image}
\alias{lmstudio_image}
\title{Submit a prompt to local LM Studio LLM}
\usage{
lmstudio_image(prompt, image_path)
}
\arguments{
\item{prompt}{Text prompt}

\item{image_path}{Path to image}
}
\description{
Submit a prompt to local LM Studio LLM
}
\examples{
\dontrun{
lmstudio_image("Describe this image","image_name.png")
}
}
