% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{ollama}
\alias{ollama}
\title{Submit a prompt to local ollama LLM}
\usage{
ollama(prompt)
}
\arguments{
\item{prompt}{Text prompt}
}
\description{
Submit a prompt to local ollama LLM
}
\examples{
\dontrun{
ollama("Qual a capital de Portugal?")
}
}
