% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openrouter.R
\name{openrouter}
\alias{openrouter}
\title{Submit a prompt to openrouter}
\usage{
openrouter(model = "openrouter/auto", prompt)
}
\arguments{
\item{model}{Model name}

\item{prompt}{Text prompt}
}
\description{
Submit a prompt to openrouter
}
\examples{
\dontrun{
openrouter("Qual a capital de Portugal?")
}
}
