% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split}
\alias{split}
\title{Split dataframe into Train and Test dataframes, returning a list with both dataframes.}
\usage{
split(df, label, p)
}
\arguments{
\item{df}{Dataframe}

\item{label}{Target variable}

\item{p}{Proportion for Train dataframe}
}
\description{
Split dataframe into Train and Test dataframes, returning a list with both dataframes.
}
\examples{
\dontrun{
result = split(dataset, dataset$label, 0.8)
}
}
