% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini.R
\name{gemini}
\alias{gemini}
\title{Submit a prompt to gemini}
\usage{
gemini(prompt)
}
\arguments{
\item{prompt}{Text prompt}
}
\description{
Submit a prompt to gemini
}
\examples{
\dontrun{
gemini("Qual a capital de Portugal?")
}
}
